refactor: 제안중, 거절된 돈길 전역 바터시트 적용하여 props 네번 내리는 구조 개선, 거절된 돈길 삭제 테스트 필요
