+++ src/components/home/proposed/ProposedDongilSection.tsx
@@ -4,10 +4,8 @@ import {
   selectProposedDongils,
   selectProposedDongilsStatus,
 } from '@store/slices/proposedDongilsSlice';
-import { Dispatch, SetStateAction } from 'react';
 import styled from 'styled-components';
 import EmptyDongil from '../EmptyDongil';
-import SkeletonDongilList from '../SkeletonDongilList';
 import ProposedDongilList from './ProposedDongilList';
 
 function ProposedDongilSection() {
