+++ src/store/slices/proposedDongilsSlice.tsx
@@ -37,11 +37,11 @@ export const approveProposedDongil = createAsyncThunk(
   'proposedDongils/approve',
   async (thunkPayload: {
     axiosPrivate: AxiosInstance;
-    idToApprove: number;
+    id: number;
     isApprove: boolean;
   }) => {
-    const { axiosPrivate, idToApprove, isApprove } = thunkPayload;
-    const response = await axiosPrivate.patch(`/challenge/${idToApprove}`, {
+    const { axiosPrivate, id, isApprove } = thunkPayload;
+    const response = await axiosPrivate.patch(`/challenge/${id}`, {
       accept: isApprove,
     });
     return response.data;
@@ -53,11 +53,11 @@ export const rejectProposedDongil = createAsyncThunk(
   'proposedDongils/reject',
   async (thunkPayload: {
     axiosPrivate: AxiosInstance;
-    idToApprove: number;
+    id: number;
     comment: string;
   }) => {
-    const { axiosPrivate, idToApprove, comment } = thunkPayload;
-    const response = await axiosPrivate.patch(`/challenge/${idToApprove}`, {
+    const { axiosPrivate, id, comment } = thunkPayload;
+    const response = await axiosPrivate.patch(`/challenge/${id}`, {
       accept: false,
       comment,
     });
